+++ pesummary/gw/file/formats/default.py
@@ -137,7 +137,8 @@ class Default(GWRead):
                     samples[num].append(f["likelihood_stats/loglr"][num])
         elif isinstance(f[path_to_samples], h5py._hl.dataset.Dataset):
             parameters = f[path_to_samples].dtype.names
-            samples = [[i[parameters.index(j)] for j in parameters] for i in f[path]]
+            samples = [[i[parameters.index(j)] for j in parameters] for i in
+                       f[path_to_samples]]
         f.close()
         injection = {i: float("nan") for i in parameters}
         return parameters, samples, injection
@@ -147,6 +148,12 @@ class Default(GWRead):
         """
         """
         return None
+
+    def add_injection_parameters_from_file(self, injection_file):
+        """
+        """
+        self.injection_parameters = self._add_injection_parameters_from_file(
+            injection_file, self._grab_injection_parameters_from_file)
 
     def add_marginalized_parameters_from_config_file(self, config_file):
         """Search the configuration file and add the marginalized parameters